groupPage on webClient

+++ UsersService/src/main/java/ru/maksimov/UsersService/controllers/GroupsController.java
@@ -7,7 +7,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import ru.maksimov.UsersService.dto.usersDto.GroupDTO;
 import ru.maksimov.UsersService.dto.usersDto.NewGroupDTO;
+import ru.maksimov.UsersService.dto.usersDto.RequestDTO;
 import ru.maksimov.UsersService.models.Group;
 import ru.maksimov.UsersService.models.RequestForMembership;
 import ru.maksimov.UsersService.models.User;
@@ -18,6 +20,7 @@ import ru.maksimov.UsersService.util.exceptions.GroupNotFoundException;
 import ru.maksimov.UsersService.util.exceptions.RequestNotFoundException;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/groups")
@@ -38,19 +41,21 @@ public class GroupsController {
     }
 
     @GetMapping
-    public ResponseEntity<List<Group>> findAll() {
-        List<Group> groups = groupsService.findAll();
+    public ResponseEntity<List<GroupDTO>> findAll() {
+        List<GroupDTO> groups = groupsService.findAll().stream()
+                .map(group -> modelMapper.map(group, GroupDTO.class))
+                .collect(Collectors.toList());
         return ResponseEntity.ok(groups);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Group> findById(@PathVariable("id") int id) {
+    public ResponseEntity<GroupDTO> findById(@PathVariable("id") int id) {
         Group group = groupsService.findById(id);
-        return ResponseEntity.ok(group);
+        return ResponseEntity.ok(modelMapper.map(group, GroupDTO.class));
     }
 
     @PostMapping
-    public ResponseEntity<Group> createGroup(@RequestParam(name = "userId") int userId,
+    public ResponseEntity<GroupDTO> createGroup(@RequestParam(name = "userId") int userId,
                                              @RequestBody @Valid NewGroupDTO newGroupDTO,
                                              BindingResult bindingResult
                                               ) {
@@ -65,12 +70,12 @@ public class GroupsController {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
         }
 
-        return ResponseEntity.ok(group);
+        return ResponseEntity.ok(modelMapper.map(group, GroupDTO.class));
     }
 
 
     @GetMapping("/{groupId}/requests")
-    public ResponseEntity<List<RequestForMembership>> getRequests(@PathVariable("groupId") int groupId,
+    public ResponseEntity<List<RequestDTO>> getRequests(@PathVariable("groupId") int groupId,
                                                                   @RequestParam(name = "visitorId") int visitorId,
                                                                   @RequestParam(name = "message", required = false) String message) {
         Group group = groupsService.findById(groupId);
@@ -78,29 +83,37 @@ public class GroupsController {
 
         if(group.getOwner().getId().equals(visitorId)){
             List<RequestForMembership> requests = requestsService.requestsForGroup(groupId);
-            return ResponseEntity.ok(requests);
+            return ResponseEntity.ok(requests.stream()
+                    .map(request -> modelMapper.map(request, RequestDTO.class))
+                    .collect(Collectors.toList()));
         }
         else {
             RequestForMembership requestForMembership = requestsService.sendRequest(group, visitor, message);
-            return ResponseEntity.ok(List.of(requestForMembership));
+            return ResponseEntity.ok(List.of(modelMapper.map(requestForMembership, RequestDTO.class)));
         }
     }
 
     @GetMapping("/{groupId}/requests/{requestId}")
-    public ResponseEntity<RequestForMembership> confirmRequest(@PathVariable("groupId") int groupId,
+    public ResponseEntity<RequestDTO> confirmRequest(@PathVariable("groupId") int groupId,
                                                                @PathVariable("requestId") int requestId,
                                                                @RequestParam(name = "visitorId") int visitorId) {
         List<RequestForMembership> requests = requestsService.requestsForGroup(groupId);
         RequestForMembership request = requestsService.findById(requestId);
         if(requests.contains(request)) {
-            Group group = request.getGroup();
-            User user = request.getUser();
-            group.addMember(user);
-            groupsService.saveGroup(group);
-            user.addGroup(group);
-            usersService.update(user.getId(), user);
-            requestsService.deleteRequest(request);
-            return ResponseEntity.ok(request);
+            if(request.getGroup().getOwner().getId().equals(visitorId)) {
+                Group group = request.getGroup();
+                User user = request.getUser();
+                group.addMember(user);
+                groupsService.saveGroup(group);
+                user.addGroup(group);
+                usersService.update(user.getId(), user);
+                requestsService.deleteRequest(request);
+                return ResponseEntity.ok(modelMapper.map(request, RequestDTO.class));
+            }
+            else {
+                // need to change
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+            }
         }
         else {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
